+++ src/lvmapi/__init__.py
@@ -1,8 +1,10 @@
 # encoding: utf-8
 
+import os
 import pathlib
 
-from sdsstools import get_package_version, read_yaml_file
+from sdsstools import get_package_version
+from sdsstools.configuration import get_config
 
 
 # pip package name
@@ -11,5 +13,7 @@ NAME = "lvmapi"
 # package name should be pip package name
 __version__ = get_package_version(path=__file__, package_name=NAME)
 
+internal_config_path = pathlib.Path(__file__).parent / "config.yaml"
+config_path = os.getenv("LVMAPI_CONFIG_PATH", None)
 
-config = read_yaml_file(pathlib.Path(__file__).parent / "config.yaml")
+config = get_config("lvmapi", config_file=internal_config_path, user_path=config_path)